agent/requirements.txt
httpx==0.28.1
selectolax==0.3.26
PyYAML==6.0.2
supabase==2.15.0
pydantic==2.11.3
python-dateutil==2.9.0.post0
